% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymaic.R
\name{polymaic}
\alias{polymaic}
\title{Performs polyMAIC.}
\usage{
polymaic(IPD, VARSTAT, TARGS, TOLS, TYPE, COEFBOUND = 5000, ID = "USUBJID")
}
\arguments{
\item{IPD}{data.frame containing IPD from index study containing matching variables}

\item{VARSTAT}{named character vector of variable (names) to match on with statistics to match on}

\item{TARGS}{named numeric vector of targets}

\item{TOLS}{named numeric vector of tolerances}

\item{TYPE}{named character vector of data type}

\item{COEFBOUND}{Parameter space limits for polynomial parameter space (e.g., (-5000, 5000)). Default value 5000.}

\item{ID}{Unique identifier}
}
\value{
Object of class 'polymaic'
}
\description{
This function takes IPD, targets, tolerances and matching variables and performs polyMAIC.
}
\examples{
## One continuous and one binary variable
TARGS_A    <- c(AGEMEAN = 55, AGESD = 6, SEXFPROP = 0.65)          ## targets
TOLS_AN    <- c(AGEMEAN = 0.005, AGESD = 0.005, SEXFPROP = 0.0005) ## tolerances (narrow)
TOLS_AW    <- c(AGEMEAN = 1, AGESD = 0.5, SEXFPROP = 0.01)         ## tolerances (wide)
VARSTAT_A  <- c(AGE = "MEAN AND SD", SEXF = "MEAN")                ## Variables and measures to match on
TYPE_A     <- c(AGE = "CON", SEXF = "BIN")                         ## Type of variables to match on

polymaic_an <- polymaic(IPD=ScenarioA, VARSTAT=VARSTAT_A, TARGS=TARGS_A, TOLS=TOLS_AN, TYPE=TYPE_A, ID="Rnum")
summary(polymaic_an)

## Three continuous and one binary variable
TARGS_C    <- c(AGEMEAN = 49.5, AGESD = 7.2, SEXFPROP = 0.683, PAINMEAN = 7.5, PAINSD = 1.5, TIMEDMEAN = 6.1, TIMEDSD=3.8)
TOLS_CN    <- c(AGEMEAN = 0.005, AGESD = 0.005, SEXFPROP = 0.0005, PAINMEAN = 0.005, PAINSD = 0.005, TIMEDMEAN = 0.01, TIMEDSD = 0.01)
TOLS_CW    <- c(AGEMEAN = 1, AGESD = 0.5, SEXFPROP = 0.01, PAINMEAN = 0.2, PAINSD = 0.1, TIMEDMEAN = 0.2, TIMEDSD = 0.2)
VARSTAT_C  <- c(AGE = "MEAN AND SD", SEXF = "MEAN", PAIN = "MEAN AND SD", TIMED = "MEAN AND SD")
TYPE_C     <- c(AGE = "CON", SEXF = "BIN", PAIN = "CON", TIMED="CON")

polymaic_cn <- polymaic(IPD=ScenarioC, VARSTAT=VARSTAT_C, TARGS=TARGS_C, TOLS=TOLS_CN, TYPE=TYPE_C, ID="Rnum")
summary(polymaic_an)

}
