% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ScenarioA}
\alias{ScenarioA}
\title{Scenario A dataset}
\format{
The individual patient data are contained in a data frame
\code{scenarioA} with 500 rows, one per individual, and 3 variables:
\describe{
\item{Rnum - Subject number}{}
\item{AGE  - Age at baseline}{}
\item{SEXF - Sex, (female = 1, male = 0)}{}
}
}
\usage{
ScenarioA
}
\description{
Secenario A dataset from from Alsop and Pont, 2022.
}
\references{
Alsop, Jonathan C., and Lawrence O. Pont. "Matching-adjusted indirect comparison via a polynomial-based non-linear optimization method." Journal of Comparative Effectiveness Research 11.8 (2022): 551-561.
}
\keyword{datasets}
